Update AppSidebar to not use anything before persistent container loaded
